move instance stats
